# openapi/spec.yaml
openapi: "3.0.3"
info:
  title: ADH6 API
  description: >
    This is the specification for **MiNET**'s ADH6 plaform. Its aim is to manage our users, devices and treasury.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
  contact:
    name: MiNET
    email: equipe@minet.net
    url: https://minet.net
servers:
  - url: https://adh6.minet.net/api
    description: ADH6 production server
  - url: https://adh6-dev.minet.net/api
    description: ADH6 staging server
  - url: https://adh6-local.minet.net/api
    description: ADH6 staging server
paths:
...
  /account:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/searchTerm"
        - description: Filters by various properties
          in: query
          name: filter
          schema:
            $ref: "#/components/schemas/AbstractAccount"
          style: deepObject
          explode: true
        - $ref: "#/components/parameters/onlyAccount"
      responses:
        "200":
          description: The list of accounts
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AbstractAccount"
                type: array
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
                example: 25
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Filter accounts
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - account
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbstractAccount"
        description: The account to create
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
          description: The created account
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Create an account
      security: 
        - OpenIdConnect:
          - treasurer:write
        - ApiKeyTreasuryAuth: []
      tags:
        - account
  /account/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      parameters:
        - $ref: "#/components/parameters/onlyAccount"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbstractAccount"
          description: The fetched account
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve an account
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - account
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbstractAccount"
        description: The new values for this account
        required: true
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Udate an account
      security: 
        - OpenIdConnect:
          - treasurer:write
        - ApiKeyTreasuryAuth: []
      tags:
      - account
  /account_type:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/searchTerm"
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AccountType"
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
                example: 25
        "400":
          description: Invalid input
      summary: Filter account types
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - account
  /account_type/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountType"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve an account type
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - account
...
  /payment_method:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/searchTerm"
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PaymentMethod"
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
                example: 25
        "400":
          description: Invalid input
      summary: Filter payment methods
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - transaction
  /payment_method/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve a payment method
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - transaction
...
  /product:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/searchTerm"
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Product"
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
                example: 25
        "400":
          description: Invalid input
      summary: Filter products
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - treasury
  /product/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve a product
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - treasury
  /product/buy:
    parameters:
      - description: Member that buy th products
        in: query
        name: memberId
        required: true
        schema:
          type: integer
      - description: List of the products to add in the transaction
        in: query
        name: products
        required: true
        schema:
          type: array
          items:
            type: integer
        explode: false
      - description: Payment Method used to buy the products
        in: query
        name: paymentMethod
        required: true
        schema:
          type: integer
    post:
      responses:
        "204":
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Generate a new transaction for a product
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
      - treasury
...
  /transaction:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/searchTerm"
        - description: Filters by various properties
          in: query
          name: filter
          schema:
            $ref: "#/components/schemas/AbstractTransaction"
          style: deepObject
          explode: true
        - $ref: "#/components/parameters/onlyTransaction"
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AbstractTransaction"
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
                example: 0
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Filter transactions
      security: 
        - OpenIdConnect:
          - treasurer:read
        - ApiKeyTreasuryAuth:
            - treasurer:read
      tags:
        - transaction
    post:
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transaction"
        description: The transaction to create
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
          description: The created transaction
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Create a transaction
      security: 
        - OpenIdConnect:
          - treasurer:write
        - ApiKeyTreasuryAuth:
            - treasurer:write
      tags:
      - transaction
  /transaction/{id}/upload:
    parameters:
      - $ref: "#/components/parameters/idPath"
    post:
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Upload an attachment to a transaction
      security: 
        - OpenIdConnect:
          - treasurer:write
        - ApiKeyTreasuryAuth:
            - treasurer:write
      tags:
        - transaction
  /transaction/{id}/validate:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      operationId: validate
      responses:
        "204":
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Validate a pending transaction
      security: 
        - OpenIdConnect:
          - treasurer:read
        - ApiKeyTreasuryAuth:
            - treasurer:read
      tags:
        - transaction
  /transaction/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      parameters:
        - $ref: "#/components/parameters/onlyTransaction"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbstractTransaction"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve a transaction
      security: 
        - OpenIdConnect:
          - treasurer:read
        - ApiKeyTreasuryAuth:
            - treasurer:read
      tags:
        - transaction
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbstractTransaction"
        description: The new values for this transaction
        required: true
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Partially update a transaction
      security: 
        - OpenIdConnect:
          - treasurer:write
        - ApiKeyTreasuryAuth:
            - treasurer:write
      tags:
        - transaction
    delete:
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Delete a transaction (MUST only be possible when pendingValidation is true)
      security: 
        - OpenIdConnect:
          - treasurer:write
        - ApiKeyTreasuryAuth:
            - treasurer:write
      tags:
        - transaction
...
components:
  parameters:
    limit:
      description: Limit the number of results returned
      in: query
      name: limit
      required: false
      schema:
        default: 25
        minimum: 0
        maximum: 100
        type: integer
    skip:
      description: Skip the first n results
      in: query
      name: offset
      required: false
      schema:
        default: 0
        minimum: 0
        type: integer
...
  securitySchemes:
    OpenIdConnect:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/auth
          tokenUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/token
          scopes:
            user: Basic user access
            admin:read: Read access to admin resources
            admin:write: Write access to admin resources
            admin:prod: Production admin access
            treasurer:read: Read access to treasury resources
            treasurer:write: Write access to treasury resources
            network:read: Read access to network resources
            network:write: Write access to network resources
            network:write:prod: Production network write access
            network:write:dev: Development network write access
            network:write:hosting: Hosting network write access
    ApiKeyAdminAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_admin_auth
    ApiKeyNetworkAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_network_auth
    ApiKeyTreasuryAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_treasury_auth
...